backend/requirements.txt
fastapi
uvicorn[standard]
pdfminer.six
python-multipart
sqlalchemy
passlib[bcrypt]
python-jose[cryptography]
python-dotenv
bcrypt
google-generativeai
joblib
numpy
scikit-learn